L8包/requirement.txt
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.7.0
beautifulsoup4==4.6.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
constantly==15.1.0
cryptography==2.3
cssselect==1.0.3
dominate==2.3.1
Flask==1.0.2
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
itsdangerous==0.24
Jinja2==2.10
lxml==4.2.3
MarkupSafe==1.0
numpy==1.15.0
parsel==1.5.0
Pillow==5.2.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
pywin32==223
queuelib==1.5.0
requests==2.19.0
scapy==2.4.0
Scrapy==1.5.1
selenium==3.13.0
service-identity==17.0.0
six==1.11.0
Twisted==18.7.0
urllib3==1.23
visitor==0.1.3
w3lib==1.19.0
Werkzeug==0.14.1
wordcloud==1.5.0
yagmail==0.11.214
zmail==0.1.8
zope.interface==4.5.0
